+++ CHANGELOG.md
@@ -11,6 +11,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `sg`: Singaporean regime
 - `tax`: `keys` for GST
 
+### Fixed
+
+- `eu-en16931-v2017`: fixed issue with payment details when due is zero
+
+### Removed
+
+- `es-verifactu-v1`: validation to prevent forbidden characters in names
+
 ### Changed
 
 - `it-sdi-v1`: Updated mapping of exemption (natura) codes to reflect CIUS mapping guide.